Partial move of graphix
* use backward compatibility imports in module
* leave python logo widget untouched
