+++ tox.ini
@@ -27,7 +27,8 @@ commands =
     urwid/widget \
     urwid/font.py \
     urwid/escape.py \
-    urwid/command_map.py
+    urwid/command_map.py \
+    urwid/graphics.py
 
 [testenv:black]
 skip_install = true
@@ -40,7 +41,8 @@ commands =
     urwid/widget \
     urwid/font.py \
     urwid/escape.py \
-    urwid/command_map.py
+    urwid/command_map.py \
+    urwid/graphics.py
 
 [testenv:ruff]
 skip_install = true
@@ -51,4 +53,5 @@ commands = ruff \
     urwid/widget \
     urwid/font.py \
     urwid/escape.py \
-    urwid/command_map.py
+    urwid/command_map.py \
+    urwid/graphics.py
